requirements.txt
Django==1.10.5
defusedxml==0.5.0
django-allauth==0.31.0
django-crispy-forms==1.6.1
oauthlib==2.0.2
pymssql==2.1.3
python3-openid==3.1.0
requests==2.13.0
requests-oauthlib==0.8.0
